Add support for file based backing store Added File store support
Added a simple file based backing store for O2.
It stores data in a 'key=value' format in the file specified.

It behaves much like a map/hash-table, in that if you add a
key that already exists, it will overwrite the old value
with the new one else a new entry is added.

The contents (i.e the value) are not encrypted and is
stored as-is.